+++ flexiapi/app/Http/Controllers/Account/VcardsStorageController.php
@@ -9,7 +9,7 @@ class VcardsStorageController extends Controller
 {
     public function index(Request $request)
     {
-        if ($this->vcardRequested($request)) {
+        //if ($this->vcardRequested($request)) {
             $vcards = '';
 
             foreach ($request->user()->vcardsStorage()->get() as $vcard) {
@@ -17,21 +17,21 @@ class VcardsStorageController extends Controller
             }
 
             return $vcards;
-        }
+        /*}
 
-        abort(404);
+        abort(404);*/
     }
 
     public function show(Request $request, string $uuid)
     {
-        return ($this->vcardRequested($request))
-            ? $request->user()->vcardsStorage()->where('uuid', $uuid)->firstOrFail()->vcard
-            : abort(404);
+        return /*($this->vcardRequested($request))
+            ?*/ $request->user()->vcardsStorage()->where('uuid', $uuid)->firstOrFail()->vcard
+            /*: abort(404)*/;
     }
 
-    private function vcardRequested(Request $request): bool
+    /*private function vcardRequested(Request $request): bool
     {
         return $request->hasHeader('content-type') == 'text/vcard'
             && $request->hasHeader('accept') == 'text/vcard';
-    }
+    }*/
 }